fix type conversion warnings-as-errors when building CMake-generated Xcode project

+++ src/mbgl/vulkan/context.cpp
@@ -438,13 +438,13 @@ bool Context::renderTileClippingMasks(gfx::RenderPass& renderPass,
         clipping.pipelineInfo.inputBindings.push_back(
             vk::VertexInputBindingDescription()
                 .setBinding(0)
-                .setStride(VertexAttribute::getStrideOf(ShaderClass::attributes[0].dataType))
+                .setStride(static_cast<uint32_t>(VertexAttribute::getStrideOf(ShaderClass::attributes[0].dataType)))
                 .setInputRate(vk::VertexInputRate::eVertex));
 
         clipping.pipelineInfo.inputAttributes.push_back(
             vk::VertexInputAttributeDescription()
                 .setBinding(0)
-                .setLocation(ShaderClass::attributes[0].index)
+                .setLocation(static_cast<uint32_t>(ShaderClass::attributes[0].index))
                 .setFormat(PipelineInfo::vulkanFormat(ShaderClass::attributes[0].dataType)));
     }
 
@@ -521,7 +521,7 @@ const vk::UniqueDescriptorSetLayout& Context::getUniformDescriptorSetLayout() {
 
         for (size_t i = 0; i < shaders::maxUBOCountPerShader; ++i) {
             bindings.push_back(vk::DescriptorSetLayoutBinding()
-                                   .setBinding(i)
+                                   .setBinding(static_cast<uint32_t>(i))
                                    .setStageFlags(stageFlags)
                                    .setDescriptorType(vk::DescriptorType::eUniformBuffer)
                                    .setDescriptorCount(1));
@@ -542,7 +542,7 @@ const vk::UniqueDescriptorSetLayout& Context::getImageDescriptorSetLayout() {
 
         for (size_t i = 0; i < shaders::maxTextureCountPerShader; ++i) {
             bindings.push_back(vk::DescriptorSetLayoutBinding()
-                                   .setBinding(i)
+                                   .setBinding(static_cast<uint32_t>(i))
                                    .setStageFlags(vk::ShaderStageFlagBits::eFragment)
                                    .setDescriptorType(vk::DescriptorType::eCombinedImageSampler)
                                    .setDescriptorCount(1));
